Support sending MTC during playback
The behavior of jumping to a position while the not playing was changed!

By clicking into a track while it is not playing it's possible to
jump to a position. When play is pressed the playback starts at that
position. The decision whether play starts at the position or from
the beginning is done inside the track. MTCSender is not informed
about the decision.

A simple solution is to always play from the beginning and use pause
to play from a position.

Disabled warning C4458 because this is not an issue in local scope.

+++ Source/TestToneGeneratorAudioSource.cpp
@@ -40,7 +40,7 @@ void TestToneGeneratorAudioSource::setEnable(int channel, bool enable)
 
 void TestToneGeneratorAudioSource::setVolume(double volume)
 {
-    level = volume;
+    level = static_cast<float>(volume);
 }
 
 double TestToneGeneratorAudioSource::getVolume() const
@@ -52,7 +52,7 @@ void TestToneGeneratorAudioSource::getNextAudioBlock(const AudioSourceChannelInf
 {
     bufferToFill.clearActiveBufferRegion();
     const ScopedLock sl(m_lock);
-    size_t numberOfChannels = bufferToFill.buffer->getNumChannels();
+    int numberOfChannels = bufferToFill.buffer->getNumChannels();
 
     switch (mode)
     {
